improved some changes

+++ src/main/resources/init_db.sql
@@ -3,7 +3,7 @@ CREATE DATABASE booksHW;
 USE booksHW;
 
 CREATE TABLE books (
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
     price INT,
     PRIMARY KEY (id)
